Enum safe_launcher::errors::LauncherError

pub enum LauncherError {
    CoreError(Box<CoreError>),
    NfsError(Box<NfsError>),
    DnsError(Box<DnsError>),
    IpcListenerCouldNotBeBound,
    IpcListenerAborted(Error),
    IpcStreamCloneError(Error),
    ReceiverChannelDisconnected,
    IpcSessionTerminated(Option<Error>),
    FailedReadingStreamPayloadSize,
    FailedWritingStreamPayloadSize,
    PathNotFound,
    InvalidPath,
    PermissionDenied,
    JsonParseError(ParserError),
    JsonDecodeError(DecoderError),
    SpecificParseError(String),
    JsonEncodeError(EncoderError),
    SymmetricDecipherFailure,
    AppAlreadyAdded,
    AppNotRegistered,
    AppActivationFailed(Error),
    ReadPayloadSizeProhibitive,
    LocalConfigAccessFailed(String),
    Unexpected(String),
}

Launcher Errors

Variants

CoreError

Error from safe_core. Boxed to hold a pointer instead of value so that this enum variant is not insanely bigger than others.

NfsError

Errors from safe_nfs

DnsError

Errors from safe_nfs

IpcListenerCouldNotBeBound

Ipc Listener could not be bound to an endpoint

IpcListenerAborted

The Ipc Listener has errored out. New apps will no longer be able to connect to Launcher

IpcStreamCloneError

The Ipc Stream could not be cloned

ReceiverChannelDisconnected

mpsc receiver has hung up

IpcSessionTerminated

IpcSession has been terminated due to either graceful shutdown or some error as indicated

FailedReadingStreamPayloadSize

Could not read the payload size from stream

FailedWritingStreamPayloadSize

Could not write the payload size to stream

PathNotFound

Unable to find/traverse directory or file path

InvalidPath

Supplied path was invalid

PermissionDenied

Permission denied - e.g. permission to access SAFEDrive etc.

JsonParseError

Could not parse payload as a valid JSON

JsonDecodeError

Could not decode valid JSON into expected Structures probably because a mandatory field was missing or a field was wrongly named etc.

SpecificParseError

JSON non-conforming to the Launcher RFC and not covered by JsonDecodeError, e.g. things like invalid base64 formatting, unreasonable/unexpected indexing, ranges etc.

JsonEncodeError

Error encoding into Json String

SymmetricDecipherFailure

Symmetric Deciphering failed for a cipher text

AppAlreadyAdded

This path to binary has already been added on this machine

AppNotRegistered

The given app is not managed by Launcher

AppActivationFailed

Starting of App as an external process has failed

ReadPayloadSizeProhibitive

Payload to read is prohibitive in size

LocalConfigAccessFailed

Unable to Read from or Write to a Local Config file.

Unexpected

Unexpected - Probably a Logic error