Function safe_launcher::ffi::drop_launcher

pub extern fn drop_launcher(launcher_handle: *const c_void)

Discard and clean up the previously allocated launcher. Use this only if the launcher is obtained from one of the client obtainment functions in this crate (create_account, log_in). Using launcher_handle after a call to this functions is undefined behaviour.