Module safe_launcher::launcher

The gateway to the Launcher core modules. It consumes safe_core::client::Client and is a packet that must be held for the duration the core library is activated.

Modules

event_data
Data that will be transferred during specific events to get a fine-grained detail about various parameters.

Structs

Launcher
A self-managed Launcher. This is a packet which will intilise and store the library state. Dropping this packet would be enough to gracefully exit the library by initiaing a domino effect via RAII.

Enums

AppHandlerEvent
This is an event subset to be used by external codes to communicate with the App Handling module. Observer registration facilities may be availed for notifications.

IpcExternalEvent
This is an event subset to be used by external codes to communicate with the IPC handling module. Observer registration facilities may be availed for notifications.