Enum safe_launcher::launcher::AppHandlerEvent

pub enum AppHandlerEvent {
    AddApp(AppDetail),
    RemoveApp(NameType),
    ActivateApp(NameType),
    ModifyAppSettings(ModifyAppSettings),
    RegisterAppAddObserver(AppHandlerObserver),
    RegisterAppRemoveObserver(AppHandlerObserver),
    RegisterAppActivateObserver(AppHandlerObserver),
    RegisterAppModifyObserver(AppHandlerObserver),
    GetAllManagedApps(Sender<Result<Vec<ManagedApp>, LauncherError>>),
    Terminate,
}

This is an event subset to be used by external codes to communicate with the App Handling module. Observer registration facilities may be availed for notifications.

Variants

AddApp

Request Launcher to add and app to manage.

RemoveApp

Request Launcher to remove a previously added app.

ActivateApp

Request Launcher to activate/start an app.

ModifyAppSettings

Modify settings for an app

RegisterAppAddObserver

Register an observer to receive notifications about status of adding of an app.

RegisterAppRemoveObserver

Register an observer to receive notifications about status of removal of an app.

RegisterAppActivateObserver

Register an observer to receive notifications about an app being activated. Note however that a successful activation does not necessarily translate into successfully managed session. It just means that app has been started. It can still crash after start, fail authentication etc. Register obeservers in IPC module to get a more fine grained information.

RegisterAppModifyObserver

Register an observer to receive notifications about status of modification of an app.

GetAllManagedApps

Obtain all apps currently being managed by Launcher.

Terminate

Gracefully exit the app handling module. After a call to this Launcher will no longer cater to any requests handled by this module. This is essentially Launcher-close scenario and Launcher must be restarted to be functional again.