Enum safe_launcher::launcher::IpcExternalEvent

pub enum IpcExternalEvent {
    GetListenerEndpoint(Sender<String>),
    EndSession(NameType),
    RegisterVerifiedSessionObserver(IpcObserver),
    RegisterUnverifiedSessionObserver(IpcObserver),
    RegisterPendingVerificationObserver(IpcObserver),
    Terminate,
    // some variants omitted
}

This is an event subset to be used by external codes to communicate with the IPC handling module. Observer registration facilities may be availed for notifications.

Variants

GetListenerEndpoint

Obtain the endpoint on which the Launcher IPC is listening to for incoming connections.

EndSession

Request IPC Server to forget a session with given app-id. The session will be terminated.

RegisterVerifiedSessionObserver

Register an observer to receive notifications about changes in verified sessions.

RegisterUnverifiedSessionObserver

Register an observer to receive notifications about changes in unverified sessions.

RegisterPendingVerificationObserver

Register an observer to receive notifications about changes in pending verifications.

Terminate

Terminate Launcher IPC - this will essentially exit all sessions and close IPC down gracefully.