Struct safe_launcher::launcher::event_data::ModifyAppSettings

pub struct ModifyAppSettings {
    pub id: NameType,
    pub name: Option<String>,
    pub local_path: Option<String>,
    pub safe_drive_access: Option<bool>,
}

Representation of settings modification data. Optional fields should be filled with some data only if it is required to modify that parameter, otherwise fill in None.

Fields

id

Unique id given to the app. This will be consistent across all machines.

name

App name for this app stored in Launcher.

local_path

Modify local binary absolute path.

safe_drive_access

If this app is allowed to have access to SAFEDrive.