Struct safe_launcher::launcher::Launcher

pub struct Launcher {
    // some fields omitted
}

A self-managed Launcher. This is a packet which will intilise and store the library state. Dropping this packet would be enough to gracefully exit the library by initiaing a domino effect via RAII.

It is intended that the main thread be the owner of this. In order for multiple threads to tap into observer registration facility all that is need to be done is use one of the getters to event senders, clone it and distribute it to other threads. These event senders will help communicate with the core library in a completely asynchronous and thread safe manner.

Methods

impl Launcher

fn new(client: Client) -> Result<Launcher, LauncherError>

Creates a new self-managed Launcher instance, which is a packet that will intilise and store the library state. Dropping this packet would be enough to gracefully exit the library by initiaing a domino effect via RAII.

fn get_ipc_event_sender(&self) -> &EventSenderToServer<IpcExternalEvent>

Event Sender to communicate with the IPC Server, for e.g. to register observers etc.

fn get_app_handler_event_sender(&self) -> &Sender<AppHandlerEvent>

Event Sender to communicate with the App Handler, for e.g. to register observers, add an app to Laucher, remove or modify an already added app, etc.