Module safe_launcher::observer

Since the entire core library is purely event driven, this module provides types and primitives that an interfacing crate (executable or another library) should make use of to easily tap into the information and events of interest happening in the core.

Modules

event_data
Data that will be transferred during specific events to get a fine-grained detail about various parameters.

Enums

AppHandlingEvent
An event subset for App Handling Event category. These are the individual events that occur within the App Handling Category umberlla.

IpcEvent
An event subset for IPC Event category. These are the individual events that occur within the IPC Category umberlla.

LauncherEventCategoy
Representation of all the event categories that the core modules will trigger for an interested observer.

Type Definitions

AppHandlerObserver
An convenience alias for registering for events/signals arising from App Handling modules.

IpcObserver
An convenience alias for registering for events/signals arising from IPC modules.

Observer
All codes interfacing with Launcher are expected to be coded in observer pattern style for collecting notifications from core that they are interested in. This is a type alias for all events that Launcher will fire to interested listeners. Interfacing codes are expected to construct these senders and register them with appropriate core modules. The registration facility will be available with the main launcher module.