Enum safe_launcher::observer::AppHandlingEvent

pub enum AppHandlingEvent {
    AppAddition(AppAddition),
    AppRemoval(AppRemoval),
    AppActivation(Result<NameType, LauncherError>),
    AppModification(AppModification),
}

An event subset for App Handling Event category. These are the individual events that occur within the App Handling Category umberlla.

Variants

AppAddition

This event is triggered when an Application has been attempted to be added into Launcher and if successful, app can now be managed from inside Launcher.

AppRemoval

This event is triggered when an Application has been attempted to be removed from Launcher and if successful, app can no longer be managed from inside Launcher for this machine.

AppActivation

This is a notification about an attempt to activate a managed application. Note however that a successful activation does not necessarily translate into successfully managed session. It just means that app has been started. It can still crash after start, fail authentication etc. Register obeservers in IPC module to get a more fine grained information.

AppModification

This event is triggered when an Application has been attempted to be modified.