Module safe_launcher::observer::event_data

Data that will be transferred during specific events to get a fine-grained detail about various parameters.

Structs

AppAddition
Data for an app that was attempted to be added to Launcher.

AppAdditionData
Data corresponding to app addition

AppModification
Data for an app that was attempted to be modified.

AppRemoval
Data for an app that was attempted to be removed from Launcher.

ModificationDetail
Optional parameters that were changed. Those marked None have not been changed and were what they were previously.

PendingVerification
Data for an app pending verification.

UnverifiedSession
Data for an unverified session.

VerifiedSession
Data for a verified session.

Enums

Action
If the object under considerating was added or removed.