Enum safe_launcher::observer::event_data::Action

pub enum Action {
    Added,
    Removed(Option<LauncherError>),
}

If the object under considerating was added or removed.

Variants

Added

Was added.

Removed

If removal was due to some internal error, the optional field will be duely filled.